backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
tensorflow-cpu==2.17.0
